// src/app/models/user.ts
export interface User {
    userId: string;
    userName: string;
    userFirstName: string;
    userVille: string;
    userEmail: string;
    userPhone: string;
    userContrat: string;
    userEmploi: string;
    userLocalisation: string;
    userDispo: string;
    userPhoto: string;
    createdAt: number;
}